アプリメタデータ関連のコマンドを1つのファイルにまとめる変更と、メタデータ更新コマンドの追加
